requirements.txt
appnope==0.1.0
backcall==0.1.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
Django==2.2.1
djangorestframework==3.9.4
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-automl==0.2.0
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
imageio==2.5.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
networkx==2.3
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
Pygments==2.4.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
requests==2.22.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.2
wcwidth==0.1.7
